# docs/specs/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: Maestro Control Plane API
  version: 3.0.0
  license:
    name: Internal Proprietary
    identifier: LicenseRef-Internal-Proprietary
  description: >-
    Maestro v3 central control-plane contract. Project scope, role and delegated
    session are derived from the authenticated server context and resource path;
    request bodies never establish authorization scope. The API does not expose
    source-branch push, protected-branch push or merge operations.
servers:
  - url: /api/v3
security:
  - oidcBearer: []
...
paths:
...
  /projects/{project_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Projects]
      operationId: getProject
      summary: Read a Project snapshot
      description: Returns the authorized Project's current persisted snapshot.
      responses:
        '200':
          description: Project
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update Project metadata or lifecycle state
      description: Applies an authorized optimistic update to Project metadata or state.
      x-maestro-permission: project.update
      x-maestro-audit-event: project.updated
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: {$ref: '#/components/schemas/ProjectUpdate'}
      responses:
        '200':
          description: Updated project
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookToken:
      type: apiKey
      in: header
      name: X-Gitlab-Token
  parameters:
...
    ProjectId:
      name: project_id
      in: path
      required: true
      description: Resource locator only; authorization scope is server-derived.
      schema: {$ref: '#/components/schemas/Identifier'}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
...
  headers:
    ETag:
      description: Strong aggregate version ETag.
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
...
    Timestamp:
      type: string
      format: date-time
    ResourceVersion:
      type: integer
      minimum: 1
...
    ProjectState:
      type: string
      enum: [draft, configuring, active, suspended, archived]
    Project:
      type: object
      additionalProperties: false
      required: [id, team_id, key, name, state, version, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        team_id: {$ref: '#/components/schemas/Identifier'}
        key: {type: string, pattern: '^[a-z][a-z0-9-]{1,31}$'}
        name: {type: string, minLength: 1, maxLength: 120}
        state: {$ref: '#/components/schemas/ProjectState'}
        version: {$ref: '#/components/schemas/ResourceVersion'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
...
    ProjectUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        state: {$ref: '#/components/schemas/ProjectState'}
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlation_id]
      properties:
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]+$'}
        message: {type: string}
        correlation_id: {type: string}
        details: {type: object, additionalProperties: true}
  responses:
...
    BadRequest:
      description: Malformed request (`INVALID_REQUEST`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication failed (`AUTHENTICATION_REQUIRED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: Authenticated principal lacks permission (`PERMISSION_DENIED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource absent or hidden by project visibility (`RESOURCE_NOT_FOUND`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: State, idempotency hash or uniqueness conflict (`STATE_CONFLICT`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    PreconditionFailed:
      description: ETag no longer matches (`VERSION_CONFLICT`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
...
    Unprocessable:
      description: Semantic validation or immutable rule failed (`VALIDATION_FAILED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    PreconditionRequired:
      description: Required conditional header missing (`PRECONDITION_REQUIRED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
...